add follows page

+++ frontend/src/hooks/useFollow.ts
@@ -1,5 +1,6 @@
 import axios from "axios";
 import { useSession } from "next-auth/react";
+import { axiosClient } from "~/services/axiosClient";
 import { Comic } from "~/types";
 
 export default function useFollow() {
@@ -8,6 +9,7 @@ export default function useFollow() {
   const userId = session?.user?.id as string;
   const add = async (mangaSlug: string) => {
     try {
+      if (!userId) return false;
       await axios.post(`/api/follow/${mangaSlug}`, {
         userId,
       });
@@ -21,6 +23,8 @@ export default function useFollow() {
 
   const get = async (mangaSlug: string) => {
     try {
+      if (!userId) return "notfollowed";
+
       const response = await axios.get(
         `/api/follow/${mangaSlug}?userId=${userId}`
       );
@@ -30,9 +34,24 @@ export default function useFollow() {
       return "notfollowed";
     }
   };
-
+  const getAllFollows = async (page: string = "1") => {
+    try {
+      if (!userId) return null;
+      const { data } = await axiosClient.get("user/follows", {
+        params: {
+          userId,
+          page,
+        },
+      });
+      return data;
+    } catch (err) {
+      console.error(err);
+      return null;
+    }
+  };
   const _delete = async (mangaSlug: string) => {
     try {
+      if (!userId) return false;
       await axios.delete(`/api/follow/${mangaSlug}?userId=${userId}`);
       return true;
     } catch (err) {
@@ -40,6 +59,18 @@ export default function useFollow() {
       return false;
     }
   };
-
-  return { add, get, _delete };
+  const setReaded = async (mangaSlug: string, lastChapterSlug: string) => {
+    try {
+      if (!userId) return false;
+      await axios.put(`/api/follow/${mangaSlug}`, {
+        userId,
+        lastChapterSlug,
+      });
+      return true;
+    } catch (err) {
+      console.error(err);
+      return false;
+    }
+  };
+  return { add, get, _delete, getAllFollows, setReaded };
 }
